Calculate respawn timings relative to an absolute position in time
Rather than decrementing a value every second, respawn times are now calculated relative to some point in time, the point when the respawn was started. This means that time will continue to decrease even over restarts rather than just indefinitely waiting at some point until the server has started or the world has been loaded.

+++ src/main/java/wtf/choco/dragoneggdrop/commands/CommandDragonRespawn.java
@@ -86,7 +86,7 @@ public final class CommandDragonRespawn implements TabExecutor {
                 return true;
             }
 
-            int respawnSeconds = (args.length >= 2) ? MathUtil.parseSeconds(args[1]) : 300; // Default 5 minutes
+            long respawnSeconds = (args.length >= 2) ? MathUtil.parseSeconds(args[1]) : 300; // Default 5 minutes
             DragonTemplate template = plugin.getDragonTemplateRegistry().getRandomTemplate();
             assert template != null; // Impossible
